Revert "Synthesize primitives witout the bytes size"
This reverts commit c7302d2343146b385e94cac6fd615b24f005a118.

It doesn't actually work, see
https://github.com/mit-plv/fiat-crypto/issues/706#issuecomment-657157869

+++ fiat-go/src/p224_32.go
@@ -33,8 +33,8 @@ import "math/bits"
  *   out1: [0x0 ~> 0xffffffff]
  */
 /*inline*/
-func fiat_p224_cmovznz_u32(out1 *uint32, arg1 uint32, arg2 uint32, arg3 uint32) {
-  var x1 uint32 = (arg1 * 0xffffffff)
+func fiat_p224_cmovznz_u32(out1 *uint32, arg1 uint8, arg2 uint32, arg3 uint32) {
+  var x1 uint32 = (uint32(arg1) * 0xffffffff)
   var x2 uint32 = ((x1 & arg3) | ((^x1) & arg2))
   *out1 = x2
 }
@@ -3014,7 +3014,7 @@ func fiat_p224_nonzero(out1 *uint32, arg1 *[7]uint32) {
  *   out1: [[0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff], [0x0 ~> 0xffffffff]]
  */
 /*inline*/
-func fiat_p224_selectznz(out1 *[7]uint32, arg1 uint32, arg2 *[7]uint32, arg3 *[7]uint32) {
+func fiat_p224_selectznz(out1 *[7]uint32, arg1 uint8, arg2 *[7]uint32, arg3 *[7]uint32) {
   var x1 uint32
   fiat_p224_cmovznz_u32(&x1, arg1, (arg2[0]), (arg3[0]))
   var x2 uint32
